(#1146) Fix some import issues in CI
